Add $1 Unistroke Recognizer
